Bad condition 2 (Best until now) (#5)
* Bad condition 2

- Use a much small and smooth command signal to enforce a lower
  excitation.

From:
```json
"command": {
  "length": 8,
  "phase": 5,
  "pattern": [2, 0, 0, 0, 0]
}
```

To:
```json
"command": {
  "length": 15,
  "phase": 10,
  "pattern": [0.2, 0.2, 0.5, 0.5, 0.8]
}
```

- Also, the logger now log information on the current real parameter.

* Hotfix: Add an update method in the main loop

This is my fault.. I forgot to add the update method, hence the results
before this commit is actually useless..

* Log the control info while looping

- Control input was processed after the simulation (when the `close`
  method is called) because of the performance. However, we realized
  that the performance is not much dependent on this logging function.
  Hence, for the convenience, we delete this hook and decide to save
  the control input information in the info dict.

* Bad condition 2: Best performance until now

The pattern of the command gets smoother than before.
```
"pattern": [0.2, 0.2, 0.5, 0.5, 0.8]
  -> [0.2, 0.2, 0.5, 0.5, 0.2, 0.2]
```

Also, the parameter is now time varying with a linear and a sine
function of the time. The results show that the RL-CMRAC outperforms
the FE-CMRAC in some cases. We need to investigate more simulation to
argue that our method is good.